removes pack reqrite
