+++ src/views/Dashboard.tsx
@@ -7,7 +7,7 @@ import {
 import { currentUser } from "@/api/auth/authSlice";
 import ForceDirectedGraphView from "@/components/shared/layout/force-directed-graph";
 import { OverlayPanel } from "@/components/shared/layout/overlay-panel";
-import PackGraphViewUpdated from "@/components/shared/layout/pack-graph-updated";
+import PackGraphView from "@/components/shared/layout/pack-graph";
 import { FindestButton } from "@/components/shared/utilities/findest-button";
 import UserAvatar from "@/components/shared/utilities/user-avatar";
 import { Button } from "@/components/ui/button";
@@ -221,11 +221,7 @@ export const Dashboard = () => {
                 </div>
               )}
               <OverlayPanel>
-                <PackGraphViewUpdated
-                  data={typesData}
-                  searchKeyword={searchKeyword}
-                  isDashBoard={true}
-                />
+                <PackGraphView data={typesData} searchKeyword={searchKeyword} isDashBoard={true} />
                 <div className="absolute inset-0 grid place-items-center rounded-sm bg-black bg-opacity-0 transition-all duration-300 ease-in-out hover:bg-opacity-50">
                   <div className="hidden text-center group-hover:block">
                     <FindestButton
