+++ src/views/DataView.tsx
@@ -1,6 +1,6 @@
 import { useGetLinkingQuery, useGetPageTypesQuery } from "@/api/activity/activityApi";
 import ForceDirectedGraphView from "@/components/shared/layout/force-directed-graph";
-import PackGraphViewUpdated from "@/components/shared/layout/pack-graph-updated";
+import PackGraphView from "@/components/shared/layout/pack-graph";
 import DataViewSearchBar from "@/components/shared/search/data-view-searchbar";
 import { FindestButton } from "@/components/shared/utilities/findest-button";
 import { useDebounceDataView } from "@/hooks/use-debounce-data-view";
@@ -165,7 +165,7 @@ export const DataView = () => {
             isDashBoard={false}
           />
         ) : (
-          <PackGraphViewUpdated data={filteredData} isDashBoard={false} />
+          <PackGraphView data={filteredData} isDashBoard={false} />
         )}
       </div>
     </motion.div>
